added recyclerviews and adapters

+++ app/src/main/java/com/devansh/contextualcards/model/Gradient.kt
@@ -3,7 +3,7 @@ package com.devansh.contextualcards.model
 import com.google.gson.annotations.SerializedName
 
 data class Gradient(
-    val angle: Int,
+    val angle: Long,
     @SerializedName("colors")
     val colorList: List<String>
 )
